Reroot classes for clear public/privilaged separation
