+++ nmdc_api_utilities/data_staging.py
@@ -4,15 +4,15 @@ import logging
 
 import requests
 
+from nmdc_api_utilities.api_client import NMDCAPIClient
 from nmdc_api_utilities.auth import NMDCAuth
 from nmdc_api_utilities.config import API_BASE_URL
 from nmdc_api_utilities.decorators import requires_auth
-from nmdc_api_utilities.nmdc_search import NMDCSearch
 
 logger = logging.getLogger(__name__)
 
 
-class JGISequencingProjectAPI(NMDCSearch):
+class JGISequencingProjectAPI(NMDCAPIClient):
     """
     Class to interact with the NMDC API to get JGI samples.
     """
@@ -177,7 +177,7 @@ class JGISequencingProjectAPI(NMDCSearch):
         return response.json()
 
 
-class JGISampleSearchAPI(NMDCSearch):
+class JGISampleSearchAPI(NMDCAPIClient):
     """
     Class to interact with the NMDC API to get JGI samples.
     """
@@ -357,7 +357,7 @@ class JGISampleSearchAPI(NMDCSearch):
         return response.json()
 
 
-class GlobusTaskAPI(NMDCSearch):
+class GlobusTaskAPI(NMDCAPIClient):
     """
     Class to interact with the NMDC API for Globus tasks.
     """
