+++ nmdc_api_utilities/metadata.py
@@ -4,15 +4,15 @@ import logging
 
 import requests
 
+from nmdc_api_utilities.api_client import NMDCAPIClient
 from nmdc_api_utilities.auth import NMDCAuth
 from nmdc_api_utilities.config import API_BASE_URL
 from nmdc_api_utilities.decorators import requires_auth
-from nmdc_api_utilities.nmdc_search import NMDCSearch
 
 logger = logging.getLogger(__name__)
 
 
-class Metadata(NMDCSearch):
+class Metadata(NMDCAPIClient):
     """
     Class to interact with the NMDC API metadata endpoints. These deal mostly with metadata management, including validation and submission.
