+++ nmdc_api_utilities/minter.py
@@ -4,15 +4,15 @@ import logging
 
 import requests
 
+from nmdc_api_utilities.api_client import NMDCAPIClient
 from nmdc_api_utilities.auth import NMDCAuth
 from nmdc_api_utilities.config import API_BASE_URL
 from nmdc_api_utilities.decorators import requires_auth
-from nmdc_api_utilities.nmdc_search import NMDCSearch
 
 logger = logging.getLogger(__name__)
 
 
-class Minter(NMDCSearch):
+class Minter(NMDCAPIClient):
     """
     Class to interact with the NMDC API to mint new identifiers.
